retrying connection if daemon already running
this fixes a case where the deamon isn't running, but 2 or more
instances are started in parallel, and only one of them spawns daemon.
The spawner then waits for the daemon running in the child process, but
the other instance sees this situation as a daemon already running in
the background. The first connection may be made after the daemon has
launched but before it has started listening.